requirements.txt
asgiref==3.4.1
Django==3.2.7
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2
django-tinymce==3.3.0
django-ckeditor==6.1.0
django-crispy-forms==1.13.0